Extract project import dialog into separate component
